# Wildcard lookup optimization