This PR improves wildcard route performance by ~10%. Instead of parsing parameters during the lookup phase, parameters are now parsed during the tree-building phase and can be directly retrieved during lookup. Previously, parsing time increased with the size of the wildcard key due to searching for the next '/', which was expensive for very long keys. This issue has been resolved.


### Benchmark Differential
````
goos: linux
goarch: amd64
pkg: github.com/tigerwill90/fox
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
                    │    old.txt    │               new.txt               │
                    │    sec/op     │   sec/op     vs base                │
StaticAll-16           11.17µ ±  2%   10.91µ ± 1%   -2.32% (p=0.000 n=10)
GithubParamsAll-16     84.64n ±  1%   75.31n ± 2%  -11.03% (p=0.000 n=10)
OverlappingRoute-16   100.60n ±  1%   87.16n ± 2%  -13.35% (p=0.000 n=10)
StaticParallel-16      9.518n ±  3%   9.234n ± 3%   -2.98% (p=0.002 n=10)
CatchAll-16            31.97n ±  2%   31.67n ± 4%        ~ (p=0.280 n=10)
CatchAllParallel-16    5.820n ± 16%   5.904n ± 7%        ~ (p=1.000 n=10)
CloneWith-16           67.27n ±  2%   66.33n ± 2%        ~ (p=0.078 n=10)
geomean                73.27n         69.96n        -4.51%
````

And comparing with Gin and Echo for wildcard route:
````
BenchmarkFox_GithubAll-16          70188             16357 ns/op               0 B/op          0 allocs/op
BenchmarkEcho_GithubAll-16         56134             20734 ns/op               0 B/op          0 allocs/op
BenchmarkGin_GithubAll-16          70112             17793 ns/op               0 B/op          0 allocs/op
````